drm: Some more refactoring
