+++ src/gamescope_shared.h
@@ -25,3 +25,19 @@ namespace gamescope
 	};
 }
 
+enum GamescopeAppTextureColorspace
+{
+	GAMESCOPE_APP_TEXTURE_COLORSPACE_LINEAR = 0,
+	GAMESCOPE_APP_TEXTURE_COLORSPACE_SRGB,
+	GAMESCOPE_APP_TEXTURE_COLORSPACE_SCRGB,
+	GAMESCOPE_APP_TEXTURE_COLORSPACE_HDR10_PQ,
+	GAMESCOPE_APP_TEXTURE_COLORSPACE_PASSTHRU,
+};
+const uint32_t GamescopeAppTextureColorspace_Bits = 3;
+
+inline bool ColorspaceIsHDR( GamescopeAppTextureColorspace colorspace )
+{
+	return colorspace == GAMESCOPE_APP_TEXTURE_COLORSPACE_SCRGB ||
+		colorspace == GAMESCOPE_APP_TEXTURE_COLORSPACE_HDR10_PQ;
+}
+
